requirements.txt
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
Django==3.2.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
httpie==2.4.0
idna==2.10
Pygments==2.9.0
PyJWT==2.1.0
PySocks==1.7.1
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
sqlparse==0.4.1
urllib3==1.26.6
